+++ CHANGELOG.md
@@ -19,7 +19,9 @@ All notable changes to this project will be documented in this file.
 - DialogHint component on oxe-web-admin
 - DialogError on oxe-web-admin and oxe-web-community
 - Default logo and favicon
-- API resources: setting/delete_favicon and setting/delete_favicon
+- API resource: setting/delete_favicon
+- API resource: setting/delete_favicon
+- API resource: public/add_public_request - in order to handle public contact forms
 - Documentation on "Settings" page on oxe-web-admin
 
 ### Fixed